+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,11 @@
+## NEXT
+
+* [swift] Fully-qualifies types in Equatable extension test.
+
+## 20.0.1
+
+* [cpp] Fixes handling of null class arguments.
+
 ## 20.0.0
 
 * Moves all codec logic to single custom codec per file.